+++ src/main/java/com/yougrocery/yougrocery/repositories/GrocerylistRepository.java
@@ -3,5 +3,5 @@ package com.yougrocery.yougrocery.repositories;
 import com.yougrocery.yougrocery.models.Grocerylist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GrocerylistRepository extends JpaRepository<Grocerylist, Long> {
+public interface GrocerylistRepository extends JpaRepository<Grocerylist, Integer> {
 }
